Fixed: permission request leave the app
Fix no running app issue when asking permissions.
Root access issue is still there.
