+++ app/screens/__init__.py
@@ -15,6 +15,8 @@ from kivy.uix.screenmanager import Screen
 from kivy.properties import BooleanProperty
 from kivy.lang import Builder
 from kivy.logger import Logger
+
+Logger.info("Screen: import coordinator")
 from coordinator import COORDINATOR
 
 Builder.load_string('''
@@ -23,6 +25,8 @@ Builder.load_string('''
         do_scroll_x: False
         do_scroll_y: False
 ''')
+
+Logger.info("Screen: MobileInsightScreenBase")
 
 class MobileInsightScreenBase(Screen):
     fullscreen = BooleanProperty(False)
@@ -49,6 +53,7 @@ class MobileInsightScreenBase(Screen):
         # raise NotImplementedError
         pass
 
+Logger.info("Screen: Importing others")
 
 from .radio import RadioScreen
 from .connectivity import ConnectivityScreen
@@ -63,6 +68,7 @@ from .help import HelpScreen
 from .privacy import PrivacyScreen
 from .plugins import PluginsScreen
 
+Logger.info("Screen: import finished")
 
 # WARNING: The ordering of the following screens should be consistent with those in mobileinsight.kv (app.go_screen(idx))
 __all__ = [
