+++ app/screens/plugins.py
@@ -14,8 +14,9 @@ from kivy.uix.gridlayout import GridLayout
 from kivy.uix.popup import Popup
 from kivy.uix.scrollview import ScrollView
 from kivy.lang import Builder
-from main import get_plugins_list
+# from main import get_plugins_list
 import main_utils
+from main_utils import get_plugins_list
 from . import MobileInsightScreenBase
 
 Builder.load_file('screens/plugins.kv')
@@ -36,15 +37,7 @@ class PluginsScreen(MobileInsightScreenBase):
     logs = deque([], MAX_LINE)
     plugins = []
     selectedPlugin = ""
-    app_list = get_plugins_list()
-    myLayout = GridLayout(cols=2, spacing=5,
-                          # orientation="vertical",
-                          size_hint_y=None,
-                          height=(len(app_list) / 2 + len(app_list) % 2) * Window.height / 4)
-    # myLayout = GridLayout()
-    popupScroll = ScrollView(size_hint_y=None, size=(Window.width, Window.height * .9))
-    popupScroll.add_widget(myLayout)
-    popup = Popup(content=popupScroll, title="Choose a plugin")
+
 
     def __init__(self, **kw):
         """
@@ -63,6 +56,15 @@ class PluginsScreen(MobileInsightScreenBase):
         super(PluginsScreen, self).__init__(**kw)
 
         self.log_viewer = None
+        self.app_list = get_plugins_list()
+        self.myLayout = GridLayout(cols=2, spacing=5,
+                              # orientation="vertical",
+                              size_hint_y=None,
+                              height=(len(self.app_list) / 2 + len(self.app_list) % 2) * Window.height / 4)
+        # myLayout = GridLayout()
+        self.popupScroll = ScrollView(size_hint_y=None, size=(Window.width, Window.height * .9))
+        self.popupScroll.add_widget(self.myLayout)
+        self.popup = Popup(content=self.popupScroll, title="Choose a plugin")
 
         self.plugins_list = get_plugins_list()
